Standardize extensibility `_placeholder` fields
These fields are needed because a Rust struct with only public fields is
treated as having an entirely public memory layout, and can be directly
constructed by a user of the library. This, however, means that we can't
add any new fields without potentially breaking code, which would
require a major semver version bump.

The usual "fix" is to add private, zero-byte fields, which prevent the
structure from being directly constructed or exhaustively matched,
allowing us to add fields as needed with no API breakage.

In the future, Rust will handle this via `#[non_exhaustive]` attributes,
but they're not yet stable.