Update custom derives to `syn` 1.0
We implement two custom derive macros:

- `#[derive(Resource)]`, which implements `trait Resource` for a type
  representing a BigML resource. This eliminates a lot of boilerplate.
- `#[derive(Updatable)]`, which takes a resource type `Foo` and
  generates a `FooUpdate` class containing only the updateable fields.

We do this using a custom derive, which is run by the compiler at
compile-time, and which walks over the source code of a `struct
SomeResource` and generates new source code.

To do this, we use the `syn` library to parse Rust source code and write
out new source code. But two things have changed recently in `syn`:

1. `syn` finally has a stable 1.0 release. This is excellent, because it
   means that all of our dependencies can now rely on a single, stable
   version of `syn`. Before, we were relying on 2 or 3 versions of
   `syn`, which substantially increased our build times.
2. The Rust grammar for metadata attributes has changed. It used to
   allow `#[name(...)]` but now it allows things like
   `#[path::to::name(...)]`. This requires some minor tweaks to our
   code.