BREAKING: Update to `tokio` 0.2
This is a fairly major update, porting `bigml` and `bigml-parallel` to
use the latest version of `tokio` and various `tokio`-based libraries
such as `reqwest`.

This involves switching to the new `futures::Stream` type, and removing
`fut.compat()` calls throughout the code base. It also involves a lot of
minor API changes.

We also make `bigml::Result<T>` a bit more clever, giving it an optional
`bigml::Result<T, E>` argument, so we no longer need to import
`std::result::Result` everywhere that we need a custom error type to
interface with a third-party API.

All of our dependencies are on current versions, and we've dropped a
number of dependencies that are now handled by `tokio`.