BREAKING: bigml-parallel: Retry failed executions
For `bigml-parallel`, we make a subtle tweak to how we retry failed
executions after a 402 Payment Required error, in order to allow a bit
more robustness on overloaded BigML accounts.

To do this, we need to keep track of which errors are likely to be
temporary.

We also mark `enum Error` and our other major enums as
`#[non_exhaustive]`, meaning that we reserve the right to add new items
later without bumping semver. We do the same for structs, replacing
`placeholder` with `#[non_exhaustive]`.