Update to latest dependencies
Before, we didn't commit `Cargo.lock`, because we used to be just a
library, and lock files aren't very useful for libraries. But now we also have a
binary, so a lockfile makes sense.

We include that, update to our latest dependencies, and update deny.toml
accordingly.

We hold `tokio-util` at 0.2.0, because our dependencies haven't upgraded
to 0.3.0 yet, and we don't want include two versions (even though it
would work).